refactor(ai-message): move isDataPart helper to lib/chat/message-parts
