+++ components/ai/ai-message.tsx
@@ -13,19 +13,6 @@ import {
   type SuggestionsPart,
 } from "@/lib/schemas/ai-data-parts";
 
-type DataPart =
-  | { type: "data-reasoning"; data: ReasoningPart }
-  | { type: "data-task"; data: TaskPart }
-  | { type: "data-suggestions"; data: SuggestionsPart };
-
-function isDataPart(p: UIMessage["parts"][number]): p is DataPart {
-  return (
-    p.type === "data-reasoning" ||
-    p.type === "data-task" ||
-    p.type === "data-suggestions"
-  );
-}
-
 export function AIMessage({
   message,
   consumedSuggestionMsgIds,
@@ -132,5 +119,3 @@ export function AIMessage({
     </div>
   );
 }
-
-export { isDataPart };
